common: audit follow-up — per-frame heap churn, parser hardening (round 3)
Third pass on top of 53e1286, picking up the remaining items from the
original audit's "other notable findings" list and a couple of fresh
finds during the sweep.

Bug fixes:
- d_sprite: D_DrawSprite was calling
    spans = malloc(sizeof(sspan_t)*MAXHEIGHT+1);
  per sprite per frame. Two issues: the `+1` is outside the
  multiplication due to operator precedence (the buffer was 1 byte
  larger, not 1 element larger), and the per-sprite malloc/free pair
  is wasteful. Replace with a static sspan_t spans[MAXHEIGHT + 1].
- common: Q_atoi accumulated into signed int. The hex path's `val << 4`
  becomes implementation-defined the moment the high bit is set, and
  the decimal `val * 10` overflows on long inputs. Switch to unsigned
  accumulation so the shift and multiply are well defined; cast to
  int at return time, preserving the historical wraparound semantics.
- net_udp, net_wins: replace
    addr->ip.l = *(int *)hostentry->h_addr_list[0];
  with memcpy. The original was a strict-aliasing violation reading a
  char * (h_addr_list[0]) through an int *. memcpy expresses the same
  intent without UB and with no codegen difference on any sane target.

Performance:
- r_edge: R_ScanEdges allocated and freed a ~96 KB basespans buffer
  every frame. Cache it in a function-local static initialized on
  first call; the buffer's size depends only on MAXSPANS / CACHE_SIZE,
  not the scene, so a one-shot allocation is correct. Same shape as
  the round-1 D_WarpScreen fix but on a much hotter buffer.
- d_sprite: see above — the malloc/free in D_DrawSprite is also gone,
  for free, since fixing the precedence bug naturally led to a static
  buffer.
- drawhulls: winding_clip and winding_split each malloc/free two small
  arrays per call. These are recursive in BSP traversal; numpoints is
  almost always < 16. Add a 128-element stack-buffer fast path with
  malloc fallback for unusually large windings. drawhulls is GLquake-
  only and not in the libretro build, but the change keeps the GL
  variant honest.
- host: Host_ServerFrame's save_/temp_host_frametime locals were float
  while the global host_frametime is double. Promote the locals to
  match, removing a float→double→float round-trip on every server
  frame subdivision step.

Cleanup:
- common: COM_filelength widened the rftell roundtrip to int64_t and
  defensively clamps to INT_MAX rather than truncating silently. The
  caller contract is unchanged (still returns int because com_filesize
  and downstream callers are int), but a >2 GiB pakfile now caps
  cleanly instead of going negative.
- common: include <limits.h> for INT_MAX.

Verified: gcc -std=gnu89 -pedantic -Wdeclaration-after-statement
-Wno-comment on all 88 common/*.c — 0 errors, same 9 cosmetic
-Wformat-truncation snprintf advisories carried over from the previous
rounds.

+++ common/host.c
@@ -576,8 +576,11 @@ _Host_ServerFrame(void)
 void
 Host_ServerFrame(void)
 {
-    float save_host_frametime;
-    float temp_host_frametime;
+    /* host_frametime is a double globally — match that here so the
+     * subdivision loop below preserves the precision of the global
+     * instead of narrowing twice per iteration. */
+    double save_host_frametime;
+    double temp_host_frametime;
 
 /* run the world state */
     pr_global_struct->frametime = host_frametime;
